add npm targets for packaging and deploying examples (#14)
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

### Issue Link, if available

### Description
Updating the integration testing workflow to create the durable
functions in our pre-production lambda environment which supports the
feature. For this to work, we need to install a custom lambda model that
supports the new `DurableConfig` property, as well as use our custom
lambda endpoint for the CLI calls. I've moved the logic for packaging
and creating the function into their own standalone npm targets in the
`examples/` directory and updated the integration test runner to call
them. I've also added a `test:integration` script at the root of the
repository that can be used to run the workflow end to end locally.

Since we're using a npm workspace, we need to package the necessary
dependencies that are needed at runtime by the examples. This isn't a
great solution and we should consider how we can optimize this part of
our build process since currently all dependencies are being installed
at the root `node_modules`.

I've also updated the integration test runner to only run the hello
world for now -- this is just temporary while the rest of the
integration test mechanism is put in place to avoid wasting compute
resources.

**Note**: I had to change the behaviour of the SDK to not fail when
`DEX_ENDPOINT` isn't present because there's a KMS exception when
setting environment variables in the pre-prod lambda stack. I need to
figure out the missing permissions with lambda stream first so we can
create functions with environment variables -- I'll sort that out in
another change.

### Demo/Screenshots
Sync Invoke is working using the example that was created in this
workflow!
```
2025-09-07T19:03:25.129Z cb07cbac-289e-46cb-8dfd-8ca969471572 INFO Hello world from a durable function!
```

### Checklist

- [X] I have filled out every section of the PR template
- [X] I have thoroughly tested this change