add script to assert example histories (#31)
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

### Issue Link, if available

### Description
Adding logic so that we assert the execution history matches the
expected history. So far this just asserts that the events have the same
"Type" in order but doesn't assert the payload or other parts of the
event.

Some of the tests are failing due to `HttpTimeout` exceptions but
unrelated to this change.