ci: update examples to run with both local and cloud test runner (#68)
*Issue #, if available:*

DAR-SDK-277

*Description of changes:*

Updating the tests for our examples to run with both the local and cloud
test runners. Currently the cloud test runners won't run in a github
action yet, but I will be updating the action to run them after this.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.