fix(sdk): handle flakiness with callback concurrency (#103)
*Issue #, if available:*
DAR-SDK-319

*Description of changes:*

The tests were failing due to callback concurrency not working properly
in some cases. The history was:
1. Callback created
2. Callback succeeded (no re-invocation)
3. Step succeeded
4. Callback not checked again properly, invocation terminates but
execution not completed.

I've updated some tests to have longer delays so that it will fail every
time if it breaks.

The TerminatingPromise class had issues where it was not resolving
properly. I updated the class to no longer extend from Promise, and just
implement it. This means `instanceof Promise` will no longer be true,
but it fixes the logic. All the methods and promise logic should work
now as expected.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.