fix(testing-sdk): stop throwing error for missing Id field in InvocationCompleted event (#104)
*Issue #, if available:*

*Description of changes:*

Integration tests are failing since the `Id` field is expected but not
found in the new `InvocationCompleted` event. This event isn't
associated with an operation so it won't have this field. For now I'm
skipping this event, but will be processing it later.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.