refactor(sdk): address bugs and simplify ExecutionContext (#117)
*Description of changes:*
- Remove customerHandlerEvent from ExecutionContext: Original Lambda
event is not useful and used in ExecutionContext and removed.
- Remove unused executionContextId from ExecutionContext: a UUID that is
not used (left over from POC)
- Remove getParentId method and simplify run-in-child-context handler
- Fix using undefined as ParentId: we were using ParentId from
ExecutionContext that didn't pave ParentId and as a result we were using
undefined
- Simplify verbose logging by removing isVerbose from ExecutionContext

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Pooya Paridel <parpooya@amazon.com>