fix(testing-sdk): update wait status when handler is not re-invoked (#157)
*Issue #, if available:*

*Description of changes:*

Fix bug in testing library where wait status was not updated when the
wait state completed, which caused the overall execution to hang.

Adding an example for this case as well.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.