feat(examples): enable some disabled tests (#163)
*Issue #, if available:*

*Description of changes:*

Enabling some example tests that were previously disabled due to missing
features or bugs that are now fixed.

There are still a few more that have not been enabled yet due to bugs
that need to be fixed in the testing library or SDK.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.