chore: export ESM output files as .mjs files (#181)
*Issue #, if available:*

*Description of changes:*

We were running into this error when importing the SDK using an
`index.mjs` file:

```
SyntaxError: Named export 'withDurableExecution' not found. The requested module '@aws/durable-execution-sdk-js'
is a CommonJS module, which may not support all module.exports as named exports.

CommonJS modules can always be imported via the default export, for example using:

import pkg from '@aws/durable-execution-sdk-js';
const { withDurableExecution } = pkg;
```

This is due to using the newer `exports` field in `package.json` instead
of the `main/module/types` field. The newer field enforces requirements
more strictly, in particular that if the imported files are `.js` files,
it will use the module type of the `package.json`, which in our case is
`CJS` since we don't have `"type": "module"` in our `package.json`.
Therefore, it tries to import our `.js` files we specified in the
`exports` field as `CJS`, which doesn't work.

The older fields do not enforce this, and will happily import `js` files
as either `CJS` or `ESM`. To fix this, I'm changing the `ESM` import to
use `.mjs` files instead of `.js` files to explicitly specify the files
as `ESM` instead of `CJS`. Our `.js` files for CJS can remain, since
those can still be imported as `CJS`.

Another option is to use `"type": "module"` and changing the `CJS`
require field to be `.cjs` instead.

TODO:
- We need to have examples that import as ESM instead of CJS to test
this thoroughly (created issue:
https://github.com/aws/aws-durable-execution-sdk-js/issues/180)

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.