chore: replace template action with postbuild script, add rollup cache to gitignore (#288)
*Issue #, if available:*

*Description of changes:*

The GH action made checks confusing since it created two checks in each
PR, and so it marked the PR as successful when it might not actually be
successful.

Also adding rollup cache to gitignore after the incremental tsc change.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.