feat(testing-sdk): add ability to wait for submitted status for waitForCallback (#299)
*Issue #, if available:*

https://github.com/aws/aws-durable-execution-sdk-js/issues/187

*Description of changes:*

Currently the statuses for operation.waitForData only has STARTED and
COMPLETED.

Submitted is useful for waiting for after a waitForCallback has been
submitted instead of just created. This allows the submitter to run to
completion instead of having the callback be instantly completed while
the submitter has not yet completed.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.