ci: add npm-publish workflow (#320)
*Issue #, if available:*

*Description of changes:* adding npm-publish Github Actions Workflow
that I tested in
https://github.com/aws/aws-durable-execution-sdk-js/pull/319. Since the
other branch is publishing an empty package, it's difficult to directly
merge from that branch. I've strictly copied over the required workflow
files. Prior to publishing our packages to npm we still need to change
the versions to 1.0.0.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.