fix(sdk): fix default types for waitForCallback to return string  (#325)
*Issue #, if available:*

*Description of changes:*

The types for waitForCallback and createCallback defaulted to `unknown`,
which makes it seem like it's an object. By default, it's actually
`string`.

Our example tests also all passed objects into the type parameter, but
all of them are actually returning strings and not objects.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.