feat(sdk): implement implicit mode-aware logging for custom loggers (#332)
*Issue #, if available:*

*Description of changes:*

For custom loggers, mode-aware logging had to be a custom
implementation. Now, it is implicit in the language SDK and custom
loggers will get this for free. The only thing that custom loggers need
to (optionally) implement is getting durable context data.

This is implemented by creating a wrapper around custom loggers which
only has `warn/info/error/debug/log` methods. Now the custom logger is
not returned directly to the user, and only our wrapper is returned.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.