chore(testing-sdk): improve exported types and jsdoc in testing library (#350)
*Issue #, if available:*

*Description of changes:*

The testing library has missing JSDoc, and exposes some internals that
are not necessary.

Also updating the README contents to be more accurate and use correct
terminology.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.