fix(testing-sdk): only use execution history in determining execution completion (#351)
*Issue #, if available:*

*Description of changes:*

Fixing issue where execution history/operation list could be out of sync
with the actual execution state. This happened since the two APIs are
not called at the same time and may not have the same data, so there can
be a race condition where the execution is completed, but the history
data is not yet populated.

Updating so that it no longer calls GetDurableExecution, and only uses
GetDurableExecutionHistory to get all execution state information.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.