fix(testing-sdk): fix duplicate start history event for retry updates (#356)
*Issue #, if available:*

*Description of changes:*

Populate execution history with correct event for step retries. There
should be a StepFailed event on each retry update if there is an `Error`
object, and `StepSucceeded` if it's not an `Error` object, not a
StepStarted event.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.