fix(sdk): fix wait-for-condition incorrect attempt counting (#357)
*Issue #, if available:*

*Description of changes:*

Fix off-by-one error in wait-for-condition. Before, the `attempt`
variable in waitStrategy would be `1` on the first attempt, then `1`
again on the second attempt, and `2` on the third attempt.

We should count the current attempt in waitStrategy based on the actual
attempt. The current attempt is equal to the last attempt from the API +
1, not the result directly from the API.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.