refactor(examples): do not require explicit name/functionName for createTests (#374)
*Issue #, if available:*

*Description of changes:*

Simplifying creating tests to not require passing an explicit name to
the function, and to automatically fetch the history file instead of
manually passing in the JSON and history events.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.