fix(testing-sdk): fix errors when running as esm (#390)
*Issue #, if available:*

https://github.com/aws/aws-durable-execution-sdk-js/issues/387

*Description of changes:*

The testing SDK does not work in ESM due to a dependency on `__dirname`
which is CJS-only.

Solution:
- Adding ESM shims to the rollup config with `@rollup/plugin-esm-shim`.
This adds `__dirname`, `__filename`, and `require` as global variables
in our bundle (only `__dirname` is used currently).
- Fix the path to the worker file when running as ESM

Additional changes:
- Remove file system checking and use compile-time variables to
determine worker path
- Add step and wait example running as ESM (in local mode using vitest,
not yet running against the cloud)


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.