fix(testing-sdk): only return new data in checkpoint response (#393)
*Issue #, if available:*

https://github.com/aws/aws-durable-execution-sdk-js/issues/384

*Description of changes:*

Currently the testing library local runner checkpoint API returns the
entire operation list on each checkpoint, instead of only what changed
since the last checkpoint call. Although this typically works correctly,
this is not correct behaviour, and can make it harder to debug issues.

This logic is also necessary to ensure that we re-invoke when there is
stale data on the language SDK side compared to the backend (i.e.
checkpoint data that has not reached the client yet).

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.