feat(testing-sdk): add full support for run-durable CLI (#392)
*Issue #, if available:*

https://github.com/aws/aws-durable-execution-sdk-js/issues/388

*Description of changes:*

run-durable CLI is currently not exported from our package, but it is
advertised.

It also uses unintuitive arguments which would not be easy to maintain.
Since it's not working right now, we can improve this before actually
launching it.

Changes:
- Change language SDK DurableExecutionInvocationInputWithClient to work
for objects that look the same, even if they are not the same instance.
This is necessary since customers might bundle the language SDK when
running their script, and we need the testing library to work with
different instances as long as they match.
- Using `commander` library as a dependency for the testing library for
argument parsing
- Change default skipTime to false for more accurate testing by default
- Refactor run-durable CLI to separate files
- Add unit and integration tests for run-durable CLI with CJS and MJS
handlers

Example usage:
```
npx run-durable --show-history step-basic.js
```

or with a custom loader

```
NODE_OPTIONS="--import tsx" npx run-durable --show-history step-basic.ts
```


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.