fix(testing-sdk): support PENDING status reinvocation and execution failure (#397)
*Issue #, if available:*

https://github.com/aws/aws-durable-execution-sdk-js/issues/384

*Description of changes:*

When the client (language SDK) has stale data, it should be reinvoked if
it exited without knowing the latest state of the checkpoint API. This
is how the actual service works, and so it should be implemented in the
testing library as well. This also fixes an issue where the language SDK
exits at the same time that a callback completes, and hangs
forever/throws an error. It will now re-invoke in this case.

Additionally, the testing SDK will now throw an error when the language
SDK returns PENDING with no outstanding pending operations, timers, or
stale data. This will fix the issue where it would hang forever when
this happened.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.