fix(sdk): handle empty items in map/parallel (#403)
*Issue #, if available:*

Closes https://github.com/aws/aws-durable-execution-sdk-js/issues/399

*Description of changes:*

The SDK does not properly handle map/parallel with an empty list. It
will just wait forever waiting for the next item.

Adding logic to resolve immediately when the lists are empty, and adding
integ/unit tests for this case.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.